Rehome some util functions
